Sicherheitsabfrage falls OpenWB bei angestecktem Auto neu gestartet wird.

+++ modules/goelp1/main.sh
@@ -75,8 +75,12 @@ if [[ $? == "0" ]] ; then
 		echo 1 > /var/www/html/openWB/ramdisk/plugstat
 		#wenn das Auto angesteckt ist, wird der ausgelesene Zählerstand ignoriert und stattdessen die Leistung aufintegriert
 		#Grund: der ausgelesene Zählerstand hat eine Auflösung von 1kWh -> zu ungenau in der Darstellung
-		simenergy=$(echo "scale=3; $(</var/www/html/openWB/ramdisk/goeposkwh)/1000" | bc)
-		echo $simenergy > /var/www/html/openWB/ramdisk/llkwh
+        if [ -f "/var/www/html/openWB/ramdisk/goeposkwh" ]; then
+		    simenergy=$(echo "scale=3; $(</var/www/html/openWB/ramdisk/goeposkwh)/1000" | bc)
+		    echo $simenergy > /var/www/html/openWB/ramdisk/llkwh
+        else
+            echo $llkwh > /var/www/html/openWB/ramdisk/llkwh
+        fi
 	fi
 	if [[ $car == "2" ]] ; then
 		echo 1 > /var/www/html/openWB/ramdisk/chargestat
